+++ TestApp/Sources/Reader/PDF/PDFModule.swift
@@ -26,8 +26,8 @@ final class PDFModule: ReaderFormatModule {
         self.delegate = delegate
     }
     
-    var publicationFormats: [Publication.Format] {
-        return [.pdf]
+    func supports(_ publication: Publication) -> Bool {
+        return publication.conforms(to: .pdf)
     }
     
     func makeReaderViewController(for publication: Publication, locator: Locator?, bookId: Book.Id, books: BookRepository, bookmarks: BookmarkRepository, resourcesServer: ResourcesServer) throws -> UIViewController {